Adjusted functionality

+++ src/model/FreeRoom.java
@@ -12,7 +12,7 @@ public class FreeRoom extends Room{
      * @param enumeration string, the hotel room type - SINGLE or DOUBLE
      */
     public FreeRoom(String roomNumber, Double price, RoomType enumeration) {
-        super(roomNumber,0.0, enumeration);
+        super(roomNumber, (double) 0, enumeration);
     }
 
     @Override
